requirements.txt
pydantic.ai>=1.2.1
python-dotenv>=1.1.1
asyncio>=4.0.0
mysql-connector-python>=9.4.0
elevenlabs>=2.20.0
pyaudio>=0.2.14
speechrecognition>=3.14.3
sounddevice>= 0.5.3
soundfile>=0.13.1